refactor(neo4j): add helper for entity conversion
